# admin/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tabio POS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /outlets/{outletId}/customers:
    get:
      operationId: getCustomers
      tags: [crm]
      summary: List customers
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [crm]
      summary: Create customer
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /outlets/{outletId}/customers/{customerId}:
    get:
      operationId: getCustomer
      tags: [crm]
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    put:
      operationId: updateCustomer
      tags: [crm]
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /outlets/{outletId}/customers/{customerId}/loyalty:
    post:
      operationId: adjustLoyaltyPoints
      tags: [crm]
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustLoyaltyRequest"
      responses:
        "200":
          description: Loyalty adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # SETTINGS
...
  /outlets/{outletId}/aggregators/zomato/orders:
    get:
      operationId: getZomatoOrders
      tags: [aggregators]
      summary: Get Zomato orders (mock)
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Zomato orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AggregatorOrder"
  /outlets/{outletId}/aggregators/zomato/orders/{aggOrderId}/accept:
    post:
      operationId: acceptZomatoOrder
      tags: [aggregators]
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: aggOrderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /outlets/{outletId}/aggregators/zomato/orders/{aggOrderId}/reject:
    post:
      operationId: rejectZomatoOrder
      tags: [aggregators]
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: aggOrderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectOrderRequest"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /outlets/{outletId}/aggregators/swiggy/orders:
    get:
      operationId: getSwiggyOrders
      tags: [aggregators]
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Swiggy orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AggregatorOrder"
  /outlets/{outletId}/aggregators/swiggy/orders/{aggOrderId}/accept:
    post:
      operationId: acceptSwiggyOrder
      tags: [aggregators]
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: aggOrderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /outlets/{outletId}/aggregators/swiggy/orders/{aggOrderId}/reject:
    post:
      operationId: rejectSwiggyOrder
      tags: [aggregators]
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: aggOrderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectOrderRequest"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /outlets/{outletId}/aggregators/menu-sync:
    post:
      operationId: syncMenuToAggregators
      tags: [aggregators]
      summary: Sync menu to all aggregators
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuSyncResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        outletId:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        loyaltyPoints:
          type: integer
        totalOrders:
          type: integer
        totalSpent:
          type: number
        lastVisit:
          type: string
          nullable: true
        birthdate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, outletId, name, phone, loyaltyPoints, totalOrders, totalSpent, createdAt]
    CreateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        birthdate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name, phone]
    AdjustLoyaltyRequest:
      type: object
      properties:
        points:
          type: integer
        reason:
          type: string
      required: [points, reason]
...
    AggregatorOrder:
      type: object
      properties:
        id:
          type: string
        platform:
          type: string
          enum: [zomato, swiggy]
        status:
          type: string
          enum: [new, accepted, preparing, ready, picked, delivered, cancelled]
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              quantity:
                type: integer
              price:
                type: number
            required: [name, quantity, price]
        subtotal:
          type: number
        totalAmount:
          type: number
        estimatedDelivery:
          type: string
        createdAt:
          type: string
      required: [id, platform, status, customerName, customerPhone, deliveryAddress, items, subtotal, totalAmount, estimatedDelivery, createdAt]
    RejectOrderRequest:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
    MenuSyncResponse:
      type: object
      properties:
        success:
          type: boolean
        zomatoSynced:
          type: boolean
        swiggySynced:
          type: boolean
        message:
          type: string
      required: [success, zomatoSynced, swiggySynced, message]