# admin/lib/api-zod/src/generated/types/menuSyncResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tabio POS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MenuSyncResponse {
  success: boolean;
  zomatoSynced: boolean;
  swiggySynced: boolean;
  message: string;
}